AppOpsManager: Fix wrong coarse location permission issue
Android intentionally defines 3 duplicated coarse location permissions in
the mapping table, but for CTA security, it does not allow one switch to
control the other 2. Therefore, it need to update the table with exact permissions.
However, it is not proper to mainline currently to conflict AOSP's way.

mydongistiny: Also fix Bluetooth Scan as coarse location
Change-Id: If534a21beb32253abb4f603c7f36b5a8b0ab0601
Signed-off-by: mydongistiny <jaysonedson@gmail.com>
Signed-off-by: Joe Maples <joe@frap129.org>